requirements.txt
djangorestframework
markdown
django-filter
langchain
langchain-ollama
langchain-chroma
pandas